cloud_runner/requirements.txt
google-cloud-storage
pandas
numpy
backtesting
vectorbt
python-dotenv
pyarrow
